requirements.txt
Faker==24.14.1
mysql_connector_repackaged==0.3.1
psutil==5.9.8
PyYAML==6.0.1
PyYAML==6.0.1
urllib3==2.2.2
